# Import/Perf: for "Use Normal Data", numpyify the code that guesses which polys are flat